add links to MR and Commit on Edit form
closes #152
